requirements.txt
streamlit
torch
torchvision
pandas
numpy
Pillow
huggingface_hub
requests
opencv-python
matplotlib
